feat #101: PortraitVC 및 WhiteNavigationVC 채택구조로 변경
